Refactor store shared content to by repository
Introduce a StoreInstallRecordRepository and new StoreSharedContentService implementation, centralizing install record management and shared-cache logic. Add FileHash helper for hash computation/verification and update callers (MrPackFileInstaller, ModpackInstanceCreationService, GameStoreContentService) to use the repository, FileHash and StorePath utilities. Rename/move many Store types into Models/Services folders and remove the old StoreSharedContentService/manifest usage in favor of StoreInstalledItems. Update tests to use the new repository and adjusted APIs.
