+++ Emerald.CoreX/Store/Models/InstalledStoreItem.cs
