+++ Emerald.CoreX/Store/Models/StoreCompatibilityResult.cs
