+++ Emerald.CoreX/Store/Models/StoreContentType.cs
