+++ Emerald.CoreX/Store/Models/StoreInstallRecord.cs
