+++ Emerald.CoreX/Store/Models/StoreTagChip.cs
