+++ Emerald.CoreX/Store/Services/IGameStoreContentService.cs
