+++ Emerald.CoreX/Store/Services/StoreFileLinkService.cs
