+++ Emerald.CoreX/Store/Services/StoreSharedContentSettingsService.cs
