+++ Emerald.CoreX/Store/Utilities/StoreDisplayFormatter.cs
