+++ Emerald/App.xaml.cs
@@ -166,6 +166,7 @@ Notes
         services.AddTransient<IModrinthStore>(provider => provider.GetRequiredService<DataPackStore>());
         services.AddTransient<IModrinthStore>(provider => provider.GetRequiredService<ModPackStore>());
         services.AddSingleton<IStoreFileLinkService, StoreFileLinkService>();
+        services.AddSingleton<IStoreInstallRecordRepository, StoreInstallRecordRepository>();
         services.AddSingleton<IStoreSharedContentSettingsService, StoreSharedContentSettingsService>();
         services.AddSingleton<IStoreSharedContentService, StoreSharedContentService>();
         services.AddTransient<IGameStoreContentService, GameStoreContentService>();
